cui: Make FillType an enum class and use more in SvxAreaTabPage
So far, both the FillType enum and integer types/indices were
used to handle the fill type, with static_cast to the enum in
some places.

Switch the enum to an enum class and use it consistently
for ButtonBox and the SvxAreaTabPage that owns it.
Stop relying on indices altogether.
Rename some methods accordingly.

Introduce a new fallback value FillType::INVALID for the
case where no fill type has been set.

This is used e.g. for the "Area" tab page in Writer's
"Format" -> "Page Style" dialog.

Change-Id: I1effe0732bd0b334b3b922f2376d2059f99eb55a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189373
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>