tdf#130857 qt weld: Hide widget scheduled for deletion
If a widget is scheduled for deletion in
QtInstanceContainer::move because it is removed
from the container, but not moved to any new one,
hide the widget in addition to calling
QObject::deleteLater. This ensures the widget
is no longer visible right away, not only when
it eventually gets deleted.
This also matchws what is already done in
QtBuilder::deleteObject.

Without this commit in place, rendering issues
due to labels of the temporary tab pages created
in SvxAreaTabPage::SetOptimalSize were seen
with SAL_VCL_QT_USE_WELDED_WIDGETS=1 with
a WIP branch for supporting Writer's "Format" ->
"Page Style" dialog, tab "Area".
(QtInstanceContainer::move gets called from the
tab page dtor, ~SfxTabPage.)

Change-Id: I8e7210d759905dfc2394af5eae6bd9d7e39e802b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189423
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins