tdf#130857 cui: Refine toggle logic in "Area" tab page
    commit 770bf5a1f42be2d7397b30b91467a7dca4dc1785
    Date:   Sun Jun 16 19:32:31 2019 +0530

        tdf#124549 Check Button toggle state before toggle.

introduced logic to toggle the state of the
toggle button in the top row of the "Area" tab page
to active again if it was toggled off by clicking on
it, because one button should always remain active,
i.e. toggling a button to inactive should only happen
by clicking another toggle button.

For the GTK and VCL implementations, the toggle signal
is only emitted when initiated by user action.
For the native Qt implementation however, signals
are (currently) also emitted when toggled programmatically,
so this logic would result in the previously active
button to still remain active when clicking on
another one.

As a consequence, multiple buttons (like "None", "Color",
"Gradient") could all be shown as active at the same
time in Writer's "Format" -> "Page Style" dialog, tab
"Area", in a WIP branch where support for that dialog
is declared with with SAL_VCL_QT_USE_WELDED_WIDGETS=1.
That doesn't make sense.

Refine the logic a bit to only reset the state to
active when the currently active button would be toggled
off without another one having been set to active
previously.

Change-Id: I36b2082a014640ef62098c80f8ab191b768ee072
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189424
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>