tdf#167721 writerfilter: adjust disabled leftChars by hangingChars
Sixth Problem - in a very inconsistent move, with hangingChars
MS Word ignores the fact that w:leftChars=0 has been disabled,
(or even never defined)
and simply treats it as a zero margin,
even if a w:left margin has been defined.

I imagine this is actually a bug in MS Word.
That bug is seen in MS Word 2003 / 2010
and still exists in MS Word 2024.

There are even weirder things happening with styles,
but when I actually apply a style to a paragraph,
then that paragraph gets different margin properties
than what is shown for the style,
so probably best not to attempt any changes to styles at all.

make CppunitTest_sw_ooxmlexport22 \
    CPPUNIT_TEST_NAME=testTdf167721_chUnits

Change-Id: I689853f631aa1cb5e144e43e75672b177fd81537
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189437
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189554