Move some more tests to sw_embedded_fonts
This move ended up nontrivial. It had unexpectedly broken already
existing testFontEmbeddingDOCX on Linux, as in

 xmltesttools.cxx:215:Assertion
 Test name: (anonymous namespace)::testFontEmbeddingDOCX::TestBody
 equality assertion failed
 - Expected: 1
 - Actual  : 0
 - In <>, XPath '/w:fonts/w:font[@w:name='DejaVu Sans']/w:embedItalic' number of nodes is incorrect

It turned out related to the oddities that I noted previously, both
in testFontEmbeddingDOCX, and in tdf166627 (see the comments about
different number of embeddings on Linux). It seems that there is a
bug specifically in DejaVu Sans font (or in the Linux font manager?),
that makes the font to have a wrong family type. It doesn't happen
with DejaVu Serif, so I changed the tests to use it, which allowed
to drop some #ifdef's.

Change-Id: Ie5a80d2cb8ee7d3b3eaf0ec5adb9a8fcd119e28f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189468
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
(cherry picked from commit 8d9b573c0a3d6c78f4d30423b27d85b6b53009cb)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189521
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>