tdf#167761 sw format redline: register the item set in the autostyle pool
Load the DOCX bugdoc, save to ODT, open it, go to the 36pt font size
format redline, click reject, should be 24th, but it's now 12pt instead.

What happens is that SwRedlineExtraData_FormatColl SfxItemSet is not
part of the autoformat mechanism, so the ODF markup can't even refer to
it. So reject works correctly after DOCX open, but not after ODF
roundtrip.

The first step here is to replace the std::unique_ptr<SfxItemSet> with a
shared pointer, similar to what SwFormatAutoFormat does for normal
character direct formatting (the redline just stores the old direct
formatting).

This allows producing the std::shared_ptr<SfxItemSet> in the autostyle
pool, so the ODF export will be able to export the old direct formatting
as a next step (not yet done here.)

Change-Id: I1f7a8083ce766bd2ec06bb8e19bfce0eaebfe405
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189492
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>