tdf#130857 qt weld: Implement QtInstancePopover::popup_at_rect
... and QtInstancePopover::popup.

See QtInstanceMenuButton::handleButtonClicked and
QtInstanceMenu::popup_at_rect for similar implementations
to show some kind of popup.

This will be used e.g. by the Calc autofilter dropdown
once support for it is declared.

For now, use QtInstance::GetNativeParentFromWeldParent to get
the native parent and don't assume that the
passed weld::Widget is a QtInstanceWidget.
For the Calc autofilter popup, the passed parent is
retrieved via Window::GetFrameWeld (see
ScGridWindow::LaunchAutoFilterMenu) which calls
SalFrame::GetFrameWeld, which is not overriden
in QtFrame yet to return a QtInstanceWindow. Doing so
would break the cases of dialogs which are not yet using
native Qt widgets (as the VCL/SalInstanceDialog implementation
requires a SalInstanceWidget parent, e.g. for proper modality),
so implementing QtFrame::GetFrameWeld needs to wait for now.

Change-Id: I9ecd04d0935944d08fb280313b80594fa588e42f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189506
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>