Drop the local mutex; rely on solar mutex locked here
The import and export code are guarded by the solar mutex. I also
think, that font list code will be guarded by that.

See https://gerrit.libreoffice.org/c/core/+/189347/comment/aa2b0bc4_7831c1b9/

This allowed to simplify the check if the font is pre-installed on
the system.

Change-Id: I29ea2b5ceca328034cc438530133a467cf9e23a9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189507
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>