lok: pdf export comments in margin
Enable print comments in margin for PDF in LOK without tiled comments.
Right now comments are only printed if they are being shown in the LO
document which excludes LOK without tiled comments. Also the post-it
manager and annotation window assume in some places that MapMode is used
and correctly set, which is not the case in LOK.

Adapted an existing test about printing comments in margin to test also
for LOK.

Change-Id: I362cb89ade6d4f53832fd9a722196d214360aae6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189511
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>