tdf#130857 qt weld: Implement QtInstanceWidget::set_stack_background
The doc in include/vcl/weld.hxx says:

    //make this widget look like a page in a notebook
    virtual void set_stack_background() = 0;

Use the background color of a QTabWidget for the color.

This will be needed to declare support for the
"File" -> "Properties" dialog in an upcoming commit.

The "Custom Properties" tab (for which the color is
explicitly set) looks OK to me with this change
in place in a WIP branch to add support for that
dialog.

Change-Id: If1467f0b2989917f7e8b7eb58bea46b45ec961f5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189520
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>