OutlinerView: Rework for standard selection visualization
The OutlinerView was one of the last using XOR selections
instead of the User-defined selection color on the resp.
system. I changed that now.

That View has no Overlay, it would have been possible to
add it and then do TextEditOnOverlay and SelectionOnOverlay
as in other situations. But there is nothing 'behind' the
text, so I decided to stay at paint, but use Primitives.

The TextContent (and 1st line Icons left) are now extracted
from the EditEngine, using the mechanisms I added recently.
That means e.g. as long as the user only scrolls, moves
cursor or zooms, it is no longer necessary to re-create
that text as Primitives. This is of course done when the
text is changed (or attributes). Also not done on seleciton
changes.

Thus in most cases it no longer does: setup the EditEngine,
set the text, format the text, layout the text and paint
the text (again using 5 layers of VCL). The 1st part is
now done on-demand when the text has changed, the 2nd part
uses an SDPR if it exists instead of OutputDevice.

The only change when visualization changes is to embed the
text primitives to the StartPoint offset & embed to a
clip primitive, and visualize transformed existing text.

This makes it much faster - try a presentation with many
slides with text and change to the EditView, scroll &
zoom around, or change selection with keyboard/mouse.

The selction uses a temporary OverlaySelection object
to keep the creation of Primitives same and equal to
other places.

The OutlineView can have multiple OutlinerView's, but all
of them - if used - are based on the same Outliner, so
this gets not too complicated, only one seq<Primitives>
is necessary.

Change-Id: Ic9cea237b6d074f4fa18a0605ed80fabd0633178
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189523
Reviewed-by: Armin Le Grand <Armin.Le.Grand@me.com>
Tested-by: Jenkins