tdf#167721 writerfilter: one more chUnit unit test
The way the right margin gets its value is slightly odd
(well perhaps not odd because I expected it to inherit a w:right,
but that doesn't ALWAYS work - in fact it rarely works)
so I wanted to make sure I had this scenario captured
in a unit test.

I was also expecting the left margin to fail (as zero),
because we have a bRightSet but not a bLeftSet.
I guess mstahl's work to separate Left Margin is working,
and not all three need to be provided if one is set.
Perhaps most of this "property spamming" code
is no longer needed...

make CppunitTest_sw_ooxmlexport22 \
    CPPUNIT_TEST_NAME=testTdf167721_chUnits4

Change-Id: I2e4a510abdbf3233da85670e40be80d1bf8e63b2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189537
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins