tdf#130857 a11y: Don't set a11y relation and name for "Custom Properties"
In the "Custom Properties" page of the "File" -> "Properties"
dialog, don't set both, an accessible name based on the text
of the label of the "column" *and* a labelled-by relation to
that label, as this is redundant.

Therefore, stop setting the labelled-by relation.

Using weld::Widget::set_accessible_relation_labeled_by
doesn't match with its intended use as described in include/vcl/weld.hxx:

    // After this call this widget is only accessibility labelled by pLabel and
    // pLabel only accessibility labels this widget
    virtual void set_accessible_relation_labeled_by(weld::Widget* pLabel) = 0;

For the "Custom Properties" page, this relation was set for each
line (i.e. for each property), i.e. if multiple properties were
added using the "Add Property" line, the
"pLabel only accessibility labels this widget" assumption
mentioned above wouldn't hold.

Looking at the a11y relations in Accerciser (with either the qt6 or gtk3
VCL plugin in use) would reveal that e.g. all the "Name" comboboxes
had the labelled-by relation to the label set, but the label only had
a single label-for relation set, i.e. this was also inconsistent.

This also simplifies supporting this dialog with native Qt widgets
using SAL_VCL_QT_USE_WELDED_WIDGETS=1 in an upcoming commit, because
QtInstanceWidget::set_accessible_relation_labeled_by hasn't been
implemented yet and would trigger an assert.
(But that still needs to be handled later for other cases.)

Change-Id: Iabd1f077d47701b90de63181c822f16c841aaf91
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189540
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>