tdf#94259: don't use the non-existed chart view during import
During the import process the chart view shapes don't exist yet
and therefore trying to infer the title size from the chart shape
can't work. Luckily for the axis titles we can reuse the title size
that was stored in the file.

There are still small differences in the positioning of various
objects between Calc and Excel but these are unrelated to this
problem.

Change-Id: Ibe243e24f7bab3eedd06ee26037bcedeac48354e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189538
Tested-by: Jenkins
Reviewed-by: Markus Mohrhard <markus.mohrhard@googlemail.com>