vcl: streamline assignment of bClipped
This clarifies the variable. The variable is reused, but the first use
is to detect whether the area *should* be clipped. The second use of the
variable determines if the area *has been* clipped. Obviously, it makes
no sense to say that the area has been clipped (true) before we actually
do the clipping, so we should be setting it to false at the start of the
if statement. Then when the clipping succeeds then we can say that it
was successful.

This also removes the need to set the variable in several else
statements and we only have to set the variable to true once the
clipping succeeds.

In the next patch, I will also remove the need for the second
modification of the bClipped variable as we can just return the bitmap -
but it is far more clear what I'm doing with this patch.

Change-Id: If8e217975a7a1fa7efff8fb3607040e3269495e5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189600
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>