tdf#167833 sw: handle more ways of pasting anchored objects
Handle it in SwTransferable::PasteOLE() for embedded objects from
non-Writer applications, SwTransferable::PasteSdrFormat() for SdrObjects
from non-Writer applications, and SwTransferable::PasteGrf() for image
files from non-LO applications.

This is a bit more difficult because some of these call
SwTransferable::PasteTargetURL() after the SwFEShell function that
creates the anchored object, and this will only work if the anchored
object is selected at that time, so the deselection must be done later
in SwTransferable.

SwTransferable::PasteTargetURL() can also be called directly and would
need a check added in that case, but i couldn't figure out what to put
on the clipboard to get there (has to be INET_IMAGE or NETSCAPE_IMAGE).

Change-Id: If068622b0998893f016b84498d0618c112460db5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189629
Reviewed-by: Michael Stahl <michael.stahl@collabora.com>
Tested-by: Jenkins