tdf#167761 sw format redline: implement ODF export
Load the bugdoc, revert the format redline, the font size should go from
36pt (new direct format) to 24pt (old direct format), but it goes to
12pt (doc default).

What happens is that we have working DOCX import to store this in
SwRedlineExtraData_FormatColl's item set, but the ODF import/export is
missing.

Add the ODT export by:
1) Adding a new SwXRedlineAutoStyle that exposes the item set, assuming
   it contains character properties, and create this in
   SwXRedlinePortion::GetPropertyValue() if the RedlineAutoFormat property
   is requested.
2) In XMLRedlineExport::ExportChangeAutoStyle(), write this item set as
   an autostyle.
3) In XMLRedlineExport::ExportChangedRegion(), refer to this autostyle.

Note that similar to delete redlines, the format redline contains the
old formatting, the new formatting is directly in the document.

Change-Id: Iedd7416e5eefc4814199d60c4a6b68dbb76136e2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189672
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins