tdf#167893 and tdf#167876 weird effects in slideshow
this was a little confusing, because the bibisect pointed at one commit,
but it was actually an earlier commit which had the bug.
It just happened that the later commit made more active use of the code
from the earlier commit.

The bug was mostly in CanvasCairoExtractBitmapData, which was, I think,
largely unused before
    commit 088a7c7c451321a800ca8d3523a18b6bb93239b7
    Author: Noel Grandin <noelgrandin@gmail.com>
    Date:   Tue Sep 24 16:18:11 2024 +0200
    remove alpha device from OutputDevice

At which point some of the optimised paths in slideshow stopped
triggering as often (which is a different issue which will need
resolving)

Also, CanvasCairoExtractBitmapData was largely wrong,
returning unpremultipled data when cairo
wants premultiplied data, and returning 4-byte RGBA when the calling
code was going to tell cairo that the format was 3-byte RGBA.

So anyhow, things now work better, and the CanvasCairoExtractBitmapData
method is considerably smaller and simpler.

Change-Id: I7d668ae08f4caf8dd0d5561a2d94e23e767216d5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189867
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins