tdf#166947 properly fix windows installation sets for aarch64
While https://learn.microsoft.com/en-us/windows/win32/msi/template-summary
says that the value should be Arm64 for aarch64, the installsets created
by the baseline fail to install/produce error 1620 on installation.
The missing bit was that for that to work properly the msi summary
information table must be at least schema version 500.

That is mentioned here:
https://learn.microsoft.com/en-us/windows/win32/msi/using-64-bit-windows-installer-packages

Thanks a lot to Mike Kaganski for that pointer

Change-Id: I5fc84f877c6af98c78107ff99920def7053a3950
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189870
Tested-by: Jenkins
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>