Don’t scroll to a comment after a keypress that doesn’t do anything
Previously whenever SidebarTxtControl sees a keypress it was calling
MakeVisible to scroll the comment into view. This happens even if the
keypress doesn’t do anything such as pressing shift. This patch makes it
instead only scroll the comment into view if PostKeyEvent returns true.

An example of a usecase where this was particularily incovenient is when
pressing Alt+Tab to look at another window. The initial alt press will
make the view scroll back to the last comment you were editing. This is
annoying if you are proofreading a long document while making
occassional comments because you end up losing your place in the
document.

Note that, at least on GTK, typing actual text into the comment doesn’t
make it scroll into view regardless of whether this patch is applied.
That seems to be because the text input is handled with IM commands.
That should probably be fixed too but I think it’s independent of this
change.

Change-Id: I324debee9ee2cc9d220a2b056362bea27da43e1c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189884
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins