cool#12608: Fix form controls invalidation.
On Online, Writer doesn't get the invalidateTiles event. Instead, it gets window invalidation event.
But form controls are rendered on the document, not on a separate window. So we need to send invalidateTiles event.
invalidateTiles event is indeed sent for simple text boxes.
Rich text boxes and date selection boxes don't get the invalidations. Seems that these controls are under an extra parent window.

This commit adds a check for form controls that are under an extra parent window. So the Online side gets the invalidation.

Signed-off-by: Gökay Şatır <gokaysatir@gmail.com>
Change-Id: I8ab9506df6d9d06247f19210c959771aeddae00d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189878
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/190798
Tested-by: Jenkins
Reviewed-by: Gökay ŞATIR <gokaysatir@collabora.com>