Hide hyperlink dialog fields based on selection
Hides the name field in all cases except writer text or draw text, as these seem to be the only types of text that can store link names.
Hides text field for non-text links, such as shapes or images.
If a shell does not set the properties m_showText or m_showName, deafult values to true, to avoid regressions for unimplemented shells.

Change-Id: I228dc7f9d65c5cb17717afb79feb515376cd4d31
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189885
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Michael Stahl <michael.stahl@collabora.com>