tdf#167761 sw format redline, char style: implement DOCX import
Open the bugdoc, there should be a format redline that tracks the change
of the char style from Strong to QuoteChar, but the doc model contains
no redlines.

What happens is that we try to handle character style as a character
property, but SwUnoCursorHelper::makeRedline() uses
SfxItemPropertySet::setPropertyValue() for CharStyleName, which calls
into SwFormatCharFormat::PutValue(), which fails, so the redline is not
created.

Fix the problem by reusing lcl_setCharStyle(), that knows how to set the
character style on an item set.

The DOCX export worked already.

Change-Id: I9479cadfdf6cb598cd4b11d49d851236c8ec09e3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189897
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>