sc: make sheet view have different header colors
So we can see when the sheet is using a sheet view. This also
mimics the UI of MSO when a sheet view is used. The color was
chosen arbitrarily. This has no effect in online os it's mostly
for debugging purpuse only.

Change-Id: Idb0a7c499ec4d87ea83ac5998d5ee7e17bd724d9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189915
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
(cherry picked from commit 3b89969c60e44110a07b12c92c538b310f17ab1b)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/191086
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>