tdf#146466: SC: Unify choices of borders between toolbar and sidebar
The preset border options available in Calc via buttons were different
between the main toolbar and the Properties sidebar, leading to a
confusing and inconsistent user experience.

As per the design team's proposal in the bug report, this patch
reworks both controls to use a single, unified 12-option grid of
the most common border presets.

This change modifies the UI definitions for both the sidebar and
toolbar controls and refactors the C++ logic to handle the new
unified layout. New icons have been created for the 'criss-cross'
style to match the various icon themes.

Change-Id: I5066a41340941684c6dc55d4e1b64d4ef995cfd8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189690
Tested-by: Jenkins
Reviewed-by: Heiko Tietze <heiko.tietze@documentfoundation.org>