tdf#168010: introduce UseTrailingEmptyLinesInLayout compat option
In PowerPoint, autoshrink text size is calculated the same way as in
Impress, ignoring the empty lines block in the end. But when the text
is positioned, these lines are taken into account, which may change
position of texts aligned to bottom.

Since this method of positioning text may change text placement very
seriously, it can't be used unconditionally, and requires a compat
flag, which is introduced here, only enabled for PPTX import, and in
ODP documents having the option explicitly set.

It is not used for other MS Office document types. I saw that Excel
behaves similarly to PowerPoint, so maybe it makes sense to enable
it for XLSX, too; on the other hand, MS Word works differently. Also
I couldn't prepare a test document in binary PPT to test hehavior.
The decisions about these file types should go in separate changes.

Commit f61ea135430d7b4a1fac7de1e57a1314fbb1b49e (editeng: use text
scaling that better mimics MSO text scaling, 2024-04-03) introduced
ImpEditEngine::FormatParagraphs, that calculated paragraphs' heights,
and returned the height of the part before trailing empty paragraphs
block. It needed that height to make decisions about text scaling;
and the implementation simply assigned zero height to the trailing
empty paragraphs, therefore adding their heights didn't count.

Now we need the heights of the trailing paragraphs in the new compat
mode. This change modifies implementation of FormatParagraphs and
CalcHeight: the latter now calculates paragraph heights regardless
of the paragraph content; and the former checks if current paragraph
height should count in the returned result. This keeps the scaling
algorithm, and at the same time, provides the required height data.
We could possibly make the height calculation depend on the compat
flag; but this would make the code more complex, without a benefit.
The actual difference that the trailing empty paragraphs do is in
SdrTextObj::impDecomposeAutoFitTextPrimitive.

Change-Id: Ie37f1d2b3393f9c52be89586c73df70b108190a1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189935
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins
(cherry picked from commit 425b55efaf1f04e8f91cb049a5820fcf61e678dd)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189945
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>