tdf#164622: introduce UseTrailingEmptyLinesInLayout compat option
In PowerPoint, autoshrink text size is calculated the same way as in
Impress, ignoring the empty lines block in the end. But when the text
is positioned, these lines are taken into account, which may change
position of texts aligned to bottom.

Since this method of positioning text may change text placement very
seriously, it can't be used unconditionally, and requires a compat
flag, which is introduced here, only enabled for PPTX import, and in
ODP documents having the option explicitly set.

It is not used for other MS Office document types. I saw that Excel
behaves similarly to PowerPoint, so maybe it makes sense to enable
it for XLSX, too; on the other hand, MS Word works differently. Also
I couldn't prepare a test document in binary PPT to test hehavior.
The decisions about these file types should go in separate changes.

Change-Id: Ie37f1d2b3393f9c52be89586c73df70b108190a1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189935
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins