tdf#148057 writerfilter: overwrite BreakType::None
Granted, this is an (extreme) edge case.
Those few things that still pass through lcl_text
instead of lcl_utext miss out on many things.

In this case, a "character" column break
was being cancelled by a paragraph NONE break
(which was there to prevent a style's page break).

Still missing is the ability to split the paragraph in half.
I assume that a page break should never be masked by a column break,
and not surprisingly no existing unit tests
exhibit any of these oddities.

make CppunitTest_sw_ooxmlexport25 \
    CPPUNIT_TEST_NAME=testTdf148057_columnBreak

Change-Id: I4a5291946c8322ee0996e593f833515ee48f5a99
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189937
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins