Make the test build always
It seems it was never built in our CI. That it needs an environment
variable to run, doesn't mean it can't build - at least to prevent
bitrot.

This adds conditions to execute setup/teardown, and specific tests.
Modernizes the string literals, and fixes the types in asserts to
enable building on Windows.

Change-Id: Ie46333b5b5e516328d7236de1ff1f3b17181c789
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189968
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins