Use o3tl::convert instead of manual conversion
The new code uses the o3tl::convert to do the conversion, as it provides
a mechanism to do the conversion among different units in LibreOffice.

The old code used 35.27777778 as the conversion factor to convert points
to 10 microns.

1 point = 1/72 inch
1 inch  = 25.4 mm = 25400 microns
factor  = 25400/(72*10) ≈ 35.27777778

Please note that for values bigger than or equal 24999964, the result
can have small difference compared to the old result. For this value,
the new conversion gives 881943174, but the old formula gives 881943175
which is slightly inaccurate, off by 1. In this case, the new conversion
using o3tl::convert() is more accurate.

Change-Id: Ifcda6f9ad2f3f4d0f623ca157edbd0b4f44041fc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/189970
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Reviewed-by: Hossein   <hossein@libreoffice.org>